BrowsingDataRemover::DataType updated to uint64_t
Chromium change:
https://chromium.googlesource.com/chromium/src/+/1706d8e3c831f5e9e309a02a0b2d602deea53c86

commit 1706d8e3c831f5e9e309a02a0b2d602deea53c86
Author: cammie <cammie@chromium.org>
Date:   Mon Apr 25 23:28:59 2022 +0000

    Shared Storage: Integrate with BrowsingDataRemover.

    We hook up with `BrowsingDataRemoverImpl` so that Shared Storage, as
    well as other Privacy Sandbox API persistent storage, will be cleared
    when the clear-site-data header is sent for storage.

    Also, we must update to using uint64_t instead of int in a couple of
    places because we add the 21st value (1 << 20) to
    `BrowsingDataRemover::DataType`.

    Moreover, we add a new
    `BrowsingDataRemover::DataType::DATA_TYPE_PRIVACY_SANDBOX` collection
    and hook this up to be cleared by the Clear-Site-Data header so that
    origin-associated storage from Privacy Sandbox APIs (including
    Shared Storage) will be cleared when "storage" is sent as a value.

    Bug: 1218540

+++ browser/browsing_data/brave_clear_browsing_data.h
@@ -6,6 +6,8 @@
 #ifndef BRAVE_BROWSER_BROWSING_DATA_BRAVE_CLEAR_BROWSING_DATA_H_
 #define BRAVE_BROWSER_BROWSING_DATA_BRAVE_CLEAR_BROWSING_DATA_H_
 
+#include <cstdint>
+
 namespace content {
 class BrowsingDataRemover;
 }
@@ -30,8 +32,8 @@ class BraveClearBrowsingData {
     // to remove data.
     virtual void BeforeClearOnExitRemoveData(
         content::BrowsingDataRemover* remover,
-        int remove_mask,
-        int origin_mask) = 0;
+        uint64_t remove_mask,
+        uint64_t origin_mask) = 0;
   };
 
  protected:
